Refactor: update ProfilePage to use Profile component, adjust route parameters, and clean up unused code in createRecipe
