+++ frontend/.react-router/types/+register.ts
@@ -11,7 +11,7 @@ type Params = {
   "/": {};
   "/auth/register": {};
   "/auth/login": {};
-  "/profile": {};
   "/recipes/create": {};
   "/ingredients/create": {};
+  "/profile": {};
 };
