+++ frontend/app/pages/auth/login.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "~/components/auth/LoginForm"
+import { LoginForm } from "~/components/auth/loginForm"
 export default function LoginPage() {
     return (
         <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
